Fixed for handling proxy packets Player movement data is now logged at world server for testing purposes

+++ src/world_server/realm_connector.cpp
@@ -455,7 +455,7 @@ namespace mmo
 
 		uint16 opCode;
 		uint32 packetSize;
-		if (!(packet >> io::read<uint16>(opCode) >> io::read<uint16>(packetSize)))
+		if (!(packet >> io::read<uint16>(opCode) >> io::read<uint32>(packetSize)))
 		{
 			return PacketParseResult::Disconnect;
 		}
